move types to types.scala
